feat(openapi): Allow to redefine default error handler
Previously it only be possible to use default `rororo` error handler in
`error_middleware`, but sometimes it is needed to redefine that, for example
when web application served both API & non-API (HTML) pages.
